format solidity contracts

+++ contracts/src/ResourceAMM.sol
@@ -6,7 +6,7 @@ import {IERC20} from "openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 import {SafeERC20} from "openzeppelin-contracts/contracts/token/ERC20/utils/SafeERC20.sol";
 import {ReentrancyGuard} from "openzeppelin-contracts/contracts/utils/ReentrancyGuard.sol";
 
-contract ResourceAMM is ERC20, ReentrancyGuard{
+contract ResourceAMM is ERC20, ReentrancyGuard {
     using SafeERC20 for IERC20;
 
     IERC20 public immutable token0;
@@ -18,41 +18,23 @@ contract ResourceAMM is ERC20, ReentrancyGuard{
     uint256 public constant FEE = 3; // 0.3%
     uint256 public constant FEE_DENOMINATOR = 1000;
 
-    event LiquidityAdded(
-        address indexed provider,
-        uint256 amount0,
-        uint256 amount1,
-        uint256 lpMinted
-    );
+    event LiquidityAdded(address indexed provider, uint256 amount0, uint256 amount1, uint256 lpMinted);
 
-    event LiquidityRemoved(
-        address indexed provider,
-        uint256 amount0,
-        uint256 amount1,
-        uint256 lpBurned
-    );
+    event LiquidityRemoved(address indexed provider, uint256 amount0, uint256 amount1, uint256 lpBurned);
 
-    event Swapped(
-        address indexed user,
-        address tokenIn,
-        uint256 amountIn,
-        uint256 amountOut
-    );
+    event Swapped(address indexed user, address tokenIn, uint256 amountIn, uint256 amountOut);
 
-    constructor(address _token0, address _token1)
-        ERC20("Resourse AMM LP", "RAMM-LP")
-    {
+    constructor(address _token0, address _token1) ERC20("Resourse AMM LP", "RAMM-LP") {
         require(_token0 != _token1, "Same Token");
 
         token0 = IERC20(_token0);
         token1 = IERC20(_token1);
     }
 
-    function addLiquidity(uint amount0, uint amount1
-    ) external nonReentrant returns (uint256 lpMinted){
+    function addLiquidity(uint256 amount0, uint256 amount1) external nonReentrant returns (uint256 lpMinted) {
         require(amount0 > 0 && amount1 > 0, "Zero Amount");
 
-        if (totalSupply() == 0){
+        if (totalSupply() == 0) {
             lpMinted = sqrt(amount0 * amount1);
         } else {
             uint256 lp0 = (amount0 * totalSupply()) / reserve0;
@@ -72,20 +54,20 @@ contract ResourceAMM is ERC20, ReentrancyGuard{
         _mint(msg.sender, lpMinted);
 
         emit LiquidityAdded(msg.sender, amount0, amount1, lpMinted);
-    }   
+    }
 
-    function removeLiquidity(
-        uint256 lpAmount,
-        uint256 minAmount0,
-        uint256 minAmount1
-    ) external nonReentrant returns (uint256 amount0, uint256 amount1) {
+    function removeLiquidity(uint256 lpAmount, uint256 minAmount0, uint256 minAmount1)
+        external
+        nonReentrant
+        returns (uint256 amount0, uint256 amount1)
+    {
         require(lpAmount > 0, "Zero LP");
 
         uint256 supply = totalSupply();
 
         amount0 = (lpAmount * reserve0) / supply;
         amount1 = (lpAmount * reserve1) / supply;
-    
+
         require(amount0 >= minAmount0, "Slippage token0");
         require(amount1 >= minAmount1, "Slippage token1");
 
@@ -96,43 +78,33 @@ contract ResourceAMM is ERC20, ReentrancyGuard{
 
         token0.safeTransfer(msg.sender, amount0);
         token1.safeTransfer(msg.sender, amount1);
-    
+
         emit LiquidityRemoved(msg.sender, amount0, amount1, lpAmount);
     }
 
-
-    function swap(address tokenIn, uint256 amountIn, uint256 minAmountOut) external nonReentrant returns (uint256 amountOut){
+    function swap(address tokenIn, uint256 amountIn, uint256 minAmountOut)
+        external
+        nonReentrant
+        returns (uint256 amountOut)
+    {
         require(amountIn > 0, "Zero input");
-        require(
-            tokenIn == address(token0) ||
-            tokenIn == address(token1),
-            "Invalid Token" 
-        );
-        
+        require(tokenIn == address(token0) || tokenIn == address(token1), "Invalid Token");
+
         bool isToken0 = tokenIn == address(token0);
 
-        (
-            IERC20 inputToken,
-            IERC20 outputToken,
-            uint256 reserveIn,
-            uint256 reserveOut
-        ) = isToken0 ? (token0, token1, reserve0, reserve1) : (token1, token0, reserve1, reserve0);
+        (IERC20 inputToken, IERC20 outputToken, uint256 reserveIn, uint256 reserveOut) =
+            isToken0 ? (token0, token1, reserve0, reserve1) : (token1, token0, reserve1, reserve0);
 
-        inputToken.safeTransferFrom(
-            msg.sender,
-            address(this),
-            amountIn
-        );
+        inputToken.safeTransferFrom(msg.sender, address(this), amountIn);
 
-        uint256 amountInWithFee = amountIn * (FEE_DENOMINATOR  - FEE);
+        uint256 amountInWithFee = amountIn * (FEE_DENOMINATOR - FEE);
 
         amountOut = (amountInWithFee * reserveOut) / ((reserveIn * FEE_DENOMINATOR) + amountInWithFee);
 
         require(amountOut >= minAmountOut, "Slippage");
         require(amountOut > 0, "Zero Output");
 
-
-        if (isToken0){
+        if (isToken0) {
             reserve0 += amountIn;
             reserve1 -= amountOut;
         } else {
@@ -142,17 +114,11 @@ contract ResourceAMM is ERC20, ReentrancyGuard{
 
         outputToken.safeTransfer(msg.sender, amountOut);
 
-        emit Swapped (msg.sender, tokenIn, amountIn, amountOut);
+        emit Swapped(msg.sender, tokenIn, amountIn, amountOut);
     }
 
-        function getAmountOut(
-        address tokenIn,
-        uint256 amountIn
-    ) external view returns (uint256 amountOut) {
-        require(
-            tokenIn == address(token0) || tokenIn == address(token1),
-            "Invalid token"
-        );
+    function getAmountOut(address tokenIn, uint256 amountIn) external view returns (uint256 amountOut) {
+        require(tokenIn == address(token0) || tokenIn == address(token1), "Invalid token");
 
         bool isToken0 = tokenIn == address(token0);
 
@@ -161,9 +127,7 @@ contract ResourceAMM is ERC20, ReentrancyGuard{
 
         uint256 amountInWithFee = amountIn * (FEE_DENOMINATOR - FEE);
 
-        amountOut =
-            (amountInWithFee * reserveOut) /
-            ((reserveIn * FEE_DENOMINATOR) + amountInWithFee);
+        amountOut = (amountInWithFee * reserveOut) / ((reserveIn * FEE_DENOMINATOR) + amountInWithFee);
     }
 
     function min(uint256 a, uint256 b) internal pure returns (uint256) {
